fixed minor bugs in getInvSubSeqIdx and getSubSeqIdx;

+++ getInvSubSeqIdx.m
@@ -4,10 +4,10 @@ if nargin < 3
     start_offset = 0;
 end
 if sub_seq_ratio < 0
-    start_idx = 1 + start_offset;
+    start_idx = int32(1 + start_offset);
     end_idx = int32(n_frames*(1 + sub_seq_ratio)) + start_offset - 1;
 else
     start_idx = int32(n_frames * sub_seq_ratio) - start_offset + 1;
-    end_idx = n_frames - start_offset;
+    end_idx = int32(n_frames - start_offset);
 end
 end
